feat(secret-provider): improve code
