+++ openaev-api/src/main/java/io/openaev/secrets/provider/SecretsProvider.java
@@ -12,6 +12,8 @@ public abstract class SecretsProvider extends BaseConnectorEntity implements Ten
 
   public static final String SERVICE_NAME = "secrets-provider";
 
+  public static final String SERVICE_NAME = "secrets-provider";
+
   @JsonProperty("secrets_provider_id")
   @Getter
   @Setter
