+++ openaev-dev/Frontend start.run.xml
@@ -2,12 +2,8 @@
   <configuration default="false" name="Frontend start" type="js.build_tools.npm">
     <package-json value="$PROJECT_DIR$/openaev-front/package.json" />
     <command value="start" />
-    <arguments value="--host" />
     <node-interpreter value="project" />
     <package-manager value="yarn" />
-    <envs>
-      <env name="__VITE_ADDITIONAL_SERVER_ALLOWED_HOSTS" value="filibook.obas.lan" />
-    </envs>
     <method v="2" />
   </configuration>
 </component>
