+++ openaev-front/src/utils/lang/fr.json
@@ -3589,7 +3589,6 @@
   "user_tags": "Tags",
   "user_teams": "Équipes",
   "Username": "Nom d'utilisateur",
-  "users": "utilisateurs",
   "Users": "Utilisateurs",
   "Users & capabilities": "Utilisateurs & capacités",
   "Uses AI. Verify results.": "Utilise AI. Vérifier les résultats.",
